enable 'tier 1' workflow tests in kbn-connector-specs
