+++ src/platform/packages/shared/kbn-workflows/common/lib/errors/invalid_yaml_syntax.ts
@@ -7,5 +7,9 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-// Re-export from @kbn/workflows — canonical location
-export { WorkflowValidationError, isWorkflowValidationError } from '@kbn/workflows';
+export class InvalidYamlSyntaxError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = 'InvalidYamlSyntaxError';
+  }
+}
