+++ src/platform/packages/shared/kbn-workflows/common/lib/errors/workflow_conflict_error.ts
@@ -1,0 +1,30 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+export class WorkflowConflictError extends Error {
+  public readonly statusCode = 409;
+
+  constructor(message: string, public readonly workflowId: string) {
+    super(message);
+    this.name = 'WorkflowConflictError';
+  }
+
+  public toJSON() {
+    return {
+      error: 'Conflict',
+      message: this.message,
+      statusCode: this.statusCode,
+      workflowId: this.workflowId,
+    };
+  }
+}
+
+export function isWorkflowConflictError(error: Error): error is WorkflowConflictError {
+  return error instanceof WorkflowConflictError;
+}
