+++ src/platform/plugins/shared/workflows_management/common/lib/regex.ts
@@ -7,95 +7,25 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export const VARIABLE_REGEX = /\{\{\s*(?<key>[\w.\s|()\[\],"']*?[\w.\s|()\[\],"'])\s*\}\}/;
-export const VARIABLE_REGEX_GLOBAL = new RegExp(VARIABLE_REGEX.source, 'g');
-export const UNFINISHED_VARIABLE_REGEX_GLOBAL =
-  /\{\{\s*(?<key>[\w.\s|()\[\],"']*?[\w.\s|()\[\],"']?)\s*$/g;
-
-export const ALLOWED_KEY_REGEX =
-  /^[a-zA-Z_$][a-zA-Z0-9_$]*(?:\.[a-zA-Z_$][a-zA-Z0-9_$]*|\[\s*(?:\d+|"[^"]*"|'[^']*')\s*\])*(?:\s*\|.*)?$/;
-
-export const PROPERTY_PATH_REGEX =
-  /^[a-zA-Z_$][a-zA-Z0-9_$]*(?:\.[a-zA-Z_$][a-zA-Z0-9_$]*|\[\s*(?:\d+|"[^"]*"|'[^']*')\s*\])*$/;
-
-// Liquid-specific regex patterns
-// Matches: {{ variable | filter_prefix (but not {{ variable | filter }})
-export const LIQUID_FILTER_REGEX = /\{\{\s*[^}]*\|\s*(\w*)\s*$/;
-
-// Matches: variable | filter_prefix within liquid blocks (outside of mustache syntax)
-/**
- * LIQUID_BLOCK_FILTER_REGEX matches a variable followed by a filter within a Liquid block, but outside mustache syntax.
- *
- * Regex breakdown:
- * (?:^|[^{])      - Non-capturing group: start of line or any character except '{' (to avoid mustache blocks)
- * \s*             - Optional whitespace
- * [^{}\s]+        - One or more characters that are not '{', '}', or whitespace (the variable name)
- * \s*             - Optional whitespace
- * \|              - Pipe character separating variable and filter
- * \s*             - Optional whitespace
- * (\w*)           - Capturing group: the filter name (alphanumeric/underscore)
- * \s*$            - Optional whitespace to end of line
- *
- * Example match: "foo | filter"
- */
-export const LIQUID_BLOCK_FILTER_REGEX = /(?:^|[^{])\s*[^{}\s]+\s*\|\s*(\w*)\s*$/;
-
-// Matches liquid keywords within a liquid block (assign, case, when, echo, etc.)
-export const LIQUID_BLOCK_KEYWORD_REGEX = /^\s*(\w*)\s*$/;
-
-// Liquid block detection patterns (global versions for matching within text)
-export const LIQUID_BLOCK_START_REGEX = /\{\%-?\s*liquid\s/g;
-export const LIQUID_BLOCK_END_REGEX = /-?\%\}/g;
-
-// Liquid template detection patterns for validation
-// Matches all Liquid expressions: {{ ... }} and {% ... %}
-export const LIQUID_EXPRESSION_REGEX_GLOBAL = /(\{\{[^}]*\}\}|\{\%[^%]*\%\})/g;
-
-// More specific patterns for different Liquid constructs
-export const LIQUID_OUTPUT_REGEX_GLOBAL = /\{\{\s*([^}]*?)\s*\}\}/g;
-export const LIQUID_TAG_REGEX_GLOBAL = /\{\%\s*([^%]*?)\s*\%\}/g;
-
-export const DYNAMIC_VALUE_REGEX = /^\$\{\{\s*\S[\s\S]*\}\}$/;
-
-/**
- * Checks if a value matches the dynamic/templated value pattern ($<something>)
- * Examples: ${{env.USER}}, ${{ref:myVar}}, ${{someVariable}}
- * Pattern: starts with ${{ and ends with }}, and any non-empty string in between
- */
-export function isDynamicValue(value: unknown): boolean {
-  if (typeof value !== 'string') {
-    return false;
-  }
-  return DYNAMIC_VALUE_REGEX.test(value);
-}
-
-export const VARIABLE_VALUE_REGEX = /^\{\{\s*\S[\s\S]*\}\}$/;
-
-/**
- * Checks if a value matches the variable pattern ({{ variable }})
- * Examples: {{ variable }}, {{ variable | filter }}
- * Pattern: starts with {{ and ends with }}, and any non-empty string in between
- */
-export function isVariableValue(value: unknown): boolean {
-  if (typeof value !== 'string') {
-    return false;
-  }
-  return VARIABLE_VALUE_REGEX.test(value);
-}
-
-// Regex to match Liquid tags: {% ... %} or {%- ... -%} (with optional dashes)
-// Matches both single-line and multi-line Liquid tag blocks
-export const LIQUID_TAG_VALUE_REGEX = /\{\%-?\s*[^%]*?\s*-?\%\}/s;
-
-/**
- * Checks if a value contains Liquid tag patterns ({% ... %} or {%- ... -%})
- * Examples: {% if condition %}, {%- if condition -%}, multi-line blocks with Liquid tags
- * Pattern: matches {% or {%- followed by content and %} or -%}
- * The 's' flag allows . to match newlines for multi-line support
- */
-export function isLiquidTagValue(value: unknown): boolean {
-  if (typeof value !== 'string') {
-    return false;
-  }
-  return LIQUID_TAG_VALUE_REGEX.test(value);
-}
+// Re-export from @kbn/workflows — canonical location
+export {
+  VARIABLE_REGEX,
+  VARIABLE_REGEX_GLOBAL,
+  UNFINISHED_VARIABLE_REGEX_GLOBAL,
+  ALLOWED_KEY_REGEX,
+  PROPERTY_PATH_REGEX,
+  LIQUID_FILTER_REGEX,
+  LIQUID_BLOCK_FILTER_REGEX,
+  LIQUID_BLOCK_KEYWORD_REGEX,
+  LIQUID_BLOCK_START_REGEX,
+  LIQUID_BLOCK_END_REGEX,
+  LIQUID_EXPRESSION_REGEX_GLOBAL,
+  LIQUID_OUTPUT_REGEX_GLOBAL,
+  LIQUID_TAG_REGEX_GLOBAL,
+  DYNAMIC_VALUE_REGEX,
+  isDynamicValue,
+  VARIABLE_VALUE_REGEX,
+  isVariableValue,
+  LIQUID_TAG_VALUE_REGEX,
+  isLiquidTagValue,
+} from '@kbn/workflows/common/lib/regex';
