+++ src/platform/plugins/shared/workflows_management/common/lib/validate_workflow_yaml.ts
@@ -7,93 +7,53 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import type { ValidateWorkflowResponseDto, WorkflowYaml } from '@kbn/workflows';
-import type { WorkflowDiagnostic } from '@kbn/workflows/types/v1';
+import type { ConnectorContractUnion, ValidateWorkflowResponseDto } from '@kbn/workflows';
+import { validateWorkflowYaml as validateWorkflowYamlBase } from '@kbn/workflows';
 import type { z } from '@kbn/zod/v4';
-import { InvalidYamlSchemaError, InvalidYamlSyntaxError } from './errors';
-import { validateLiquidTemplate } from './validate_liquid_template';
-import { validateStepNameUniqueness } from './validate_step_names';
 import type { TriggerDefinitionForValidateTriggers } from './validate_triggers';
 import { validateTriggers } from './validate_triggers';
-import { parseWorkflowYamlToJSON } from './yaml';
 
 export interface ValidateWorkflowYamlOptions {
   triggerDefinitions?: TriggerDefinitionForValidateTriggers[];
+  getConnectors?: () => ConnectorContractUnion[];
 }
 
+/**
+ * Validates workflow YAML with full plugin-level validation including trigger validation.
+ * Delegates core validation (schema, liquid, step names) to @kbn/workflows and adds
+ * trigger-specific validation from workflows_extensions.
+ */
 export function validateWorkflowYaml(
   yaml: string,
   zodSchema: z.ZodType,
   options?: ValidateWorkflowYamlOptions
 ): ValidateWorkflowResponseDto {
-  const diagnostics: WorkflowDiagnostic[] = [];
-  let parsedWorkflow: WorkflowYaml | undefined;
+  // Delegate core validation to @kbn/workflows
+  const baseResult = validateWorkflowYamlBase(yaml, zodSchema, {
+    getConnectors: options?.getConnectors,
+  });
 
-  const parseResult = parseWorkflowYamlToJSON(yaml, zodSchema);
-
-  if (!parseResult.success) {
-    const { error } = parseResult;
-
-    if (error instanceof InvalidYamlSyntaxError) {
-      diagnostics.push({ severity: 'error', message: error.message, source: 'yaml-syntax' });
-    } else if (error instanceof InvalidYamlSchemaError) {
-      if (error.formattedZodError?.issues) {
-        for (const issue of error.formattedZodError.issues) {
-          diagnostics.push({
-            severity: 'error',
-            message: issue.message,
-            source: 'schema',
-            path: issue.path as (string | number)[],
-          });
-        }
-      } else {
-        diagnostics.push({ severity: 'error', message: error.message, source: 'schema' });
-      }
-    } else {
-      diagnostics.push({ severity: 'error', message: error.message, source: 'yaml-syntax' });
-    }
-  }
-
-  if (parseResult.success) {
-    parsedWorkflow = parseResult.data as unknown as WorkflowYaml;
-
-    try {
-      const stepValidation = validateStepNameUniqueness(parsedWorkflow);
-      for (const stepError of stepValidation.errors) {
-        diagnostics.push({
-          severity: 'error',
-          message: stepError.message,
-          source: 'step-name',
-        });
-      }
-    } catch {
-      // Structure too malformed for step name validation
+  // Add trigger validation if trigger definitions are provided (plugin-specific)
+  if (baseResult.parsedWorkflow && options?.triggerDefinitions) {
+    const triggerValidation = validateTriggers(
+      baseResult.parsedWorkflow,
+      options.triggerDefinitions
+    );
+    for (const triggerError of triggerValidation.errors) {
+      baseResult.diagnostics.push({
+        severity: 'error',
+        message: triggerError.message,
+        source: 'trigger',
+      });
     }
 
-    if (options?.triggerDefinitions) {
-      const triggerValidation = validateTriggers(parsedWorkflow, options.triggerDefinitions);
-      for (const triggerError of triggerValidation.errors) {
-        diagnostics.push({
-          severity: 'error',
-          message: triggerError.message,
-          source: 'trigger',
-        });
-      }
-    }
+    // Recompute valid status after adding trigger errors
+    const hasErrors = baseResult.diagnostics.some((d) => d.severity === 'error');
+    return {
+      ...baseResult,
+      valid: !hasErrors,
+    };
   }
 
-  const liquidErrors = validateLiquidTemplate(yaml, parseResult.document);
-  for (const liquidError of liquidErrors) {
-    diagnostics.push({
-      severity: 'error',
-      message: liquidError.message,
-      source: 'liquid',
-    });
-  }
-
-  return {
-    valid: diagnostics.filter((d) => d.severity === 'error').length === 0,
-    diagnostics,
-    parsedWorkflow,
-  };
+  return baseResult;
 }
